requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.8
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.17.0
findpython==0.6.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
packaging==24.2
pbs-installer==2025.2.12
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.1.1
poetry-core==2.1.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.0
tomlkit==0.13.2
trove-classifiers==2025.3.3.18
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets==15.0
zstandard==0.23.0
